+++ changelog.md
@@ -10,6 +10,8 @@
     * Regions Unexplored
     * Nature's Spirit
 * Mirror world prologue spawn - you will get teleported to the ruined world once you die or sleep
+* Tier 1 sapling recipes and quests
+* Way to craft lapis sulfur
 
 
 ## Removed
@@ -17,8 +19,13 @@
 * Rats
 * Tree Tap
 
+
 ## Fixed 
 * Incorrect requirements for offering altar ritual
 * Changed carry on to use whitelist for blocks instead of blacklist
 * Disabled clay buckets since they're annoying
 
+
+## Changed
+* Enchanted altar radius to 32 blocks
+* Theurgy digestion much faster